test: satisfy LocalAI CLI coverage lint ratchets
